refactor(trip): rename TripElement sealed classes and improve logic
- Renames `TripElement.Act` to `TripElement.TripActivity`.
- Renames `TripElement.Segment` to `TripElement.TripSegment`.
- Updates the data types for `distanceMeter` and `durationMinutes` in `RouteSegment` from `Double` to `Int`.
- Modifies `getUpcomingTripElements` in the `Trip` class to include a boolean to optionally include the currently ongoing element in the returned list.
- Updates the `estimatedTime()` function in the `Activity` class to return an `Int`.

+++ app/src/main/java/com/github/swent/swisstravel/model/trip/Activity.kt
@@ -24,7 +24,7 @@ data class Activity(
    *
    * @return The duration of the activity in minutes.
    */
-  fun estimatedTime(): Long {
-    return (endDate.seconds - startDate.seconds) / 60
+  fun estimatedTime(): Int {
+    return ((endDate.seconds - startDate.seconds) / 60).toInt()
   }
 }
